refactor: 리뷰 반영 (테스트코드 수정 중)
- Point 등록일자 삭제
- point amount Long -> Integer
- test code 구현 중

+++ point-api/point-application/src/main/java/com/flab/point/application/ChargePointProcessor.java
@@ -12,7 +12,7 @@ public class ChargePointProcessor {
         this.pointRepository = pointRepository;
     }
 
-    public Long execute(Long userId, ChargePointCommand command) {
+    public Integer execute(Long userId, ChargePointCommand command) {
         var point = pointRepository.findByUserId(userId);
         return point.add(command.getChargeAmount(), PointCategory.CHARGE);
     }
